기능(컴포넌트): 버튼 컴포넌트 리팩토링 및 e2e 테스트 추가
버튼 컴포넌트의 재사용성을 높이고 엔드투엔드 테스트를 추가했습니다.

- page.js 파일에서 HTML 버튼을 재사용 가능한 Button 컴포넌트로 교체했습니다.
- 버튼 컴포넌트용 CSS 모듈 스타일을 추가했습니다.
- 버튼, 커뮤니티, 공지사항, 작업, 학습, 프로필 페이지에 대한 e2e 테스트를 추가했습니다.
- 단위 테스트를 위한 Jest 설정을 완료했습니다.
- e2e 테스트를 위한 Playwright 설정을 완료했습니다.
- 컴포넌트 스타일을 CSS 모듈을 사용하도록 리팩토링했습니다.
- 버튼, 커뮤니티 등의 테스트 파일을 생성했습니다.
- package.json에 테스트 스크립트를 업데이트했습니다.
- 컨텐츠 생성 시 타임스탬프를 추가하여 데이터 중복 문제를 해결했습니다.

+++ provider/app/(common)/profile/delete/DeleteAccount.jsx
@@ -1,6 +1,7 @@
 "use client";
 
 import { deleteAccount } from "@/app/actions/account";
+import Button from "@/components/common/Button";
 import { useRouter } from "next/navigation";
 import { useFormStatus } from "react-dom";
 
@@ -8,12 +9,9 @@ const SubmitButton = () => {
   const { pending } = useFormStatus();
 
   return (
-    <button
-      className="button-L-2 background-purple-01 h5-16 color-white"
-      disabled={pending}
-    >
+    <Button variant="primary" size="large" type="submit" disabled={pending}>
       {pending ? "탈퇴 처리중입니다..." : "탈퇴신청"}
-    </button>
+    </Button>
   );
 };
 
@@ -22,6 +20,10 @@ export default function DeleteAccount() {
 
   const handleDeleteAccount = async (e) => {
     try {
+      const isDeleted = confirm("탈퇴를 진행하시겠습니까?");
+
+      if (!isDeleted) return;
+
       await deleteAccount();
       alert("탈퇴 신청이 완료되었습니다.");
